application_server_requirement.txt
docopt==0.6.2
pymongo==3.7.2
stomp.py==4.1.21